requirements.txt
python-telegram-bot~=21.7
openai~=1.54.3
environs~=11.0.0
google-api-python-client~=2.151.0
google-auth-httplib2
google-auth-oauthlib~=1.2.1
protobuf~=5.28.3